+++ Sources/KeyboardKit/Gestures/KeyboardGestures.swift
@@ -82,7 +82,7 @@ struct KeyboardGestures<Content: View>: View {
     var body: some View {
         view.overlay(
             GeometryReader { geo in
-                if #available(iOS 14.0, macOS 11.0, watchOS 8.0, *) {
+                if isNewGestureEngineEnabled, #available(iOS 14.0, macOS 11.0, watchOS 8.0, *) {
                     GestureButton(
                         isPressed: isPressed,
                         pressAction: { handlePress(in: geo) },
@@ -117,6 +117,12 @@ private extension KeyboardGestures {
     var actionCalloutContext: ActionCalloutContext? { .shared }
 
     var inputCalloutContext: InputCalloutContext? { .shared }
+
+    var featureToggle: FeatureToggle { .shared }
+
+    var isNewGestureEngineEnabled: Bool {
+        featureToggle.isFeatureEnabled(.newButtonGestureEngine)
+    }
 }